requirements.txt
fastapi
starkbank
uvicorn[standard]
python-dotenv
google-cloud-secret-manager
pytest
apscheduler
httpx